requirements.txt
colorama~=0.4.6
efficientnet_pytorch~=0.7.1
matplotlib~=3.9.2
numba~=0.60.0
numpy~=2.1.1
opencv_python~=4.10.0.84
Pillow~=10.4.0
scikit_learn~=1.5.2
seaborn~=0.13.2
tabulate~=0.9.0
torch~=2.4.0+cu124
torch~=2.4.1+cu124
torch_optimizer~=0.3.0
torchcam~=0.4.0
torchsummary~=1.5.1
torchvision~=0.19.0+cu124
torchvision~=0.19.1+cu124
tqdm~=4.66.5
vit_pytorch~=1.7.12
